refactored acylic_graph to include agraph class, refactored changes among c++ project
