+++ src/acyclic_graph.cpp
@@ -7,194 +7,346 @@
  * This file contains the functions associated with an acyclic graph
  * representation of a symbolic equation.
  */
-
-#include <stdlib.h>
-#include <stdio.h>
-#include <math.h>
 #include <iostream>
-
-#include <Eigen/Dense>
+#include <iomanip>
 
 #include "BingoCpp/acyclic_graph.h"
-#include "BingoCpp/graph_manip.h"
-#include "BingoCpp/backend_nodes.h"
+#include "BingoCpp/backend.h"
+#include "BingoCpp/acyclic_graph_nodes.h"
 
+OperatorInterface oper_interface;
 
-const int NODE_IDX = 0;
-const int OP_1 = 1;
-const int OP_2 = 2;
-
-bool IsCpp() {
-    return true;
+AcyclicGraph::AcyclicGraph() {
+  stack = Eigen::ArrayX3i(0, 3);
+  constants = Eigen::VectorXd(0);
+  simple_stack = Eigen::ArrayX3i(0, 3);
+  fitness = std::vector<double>();
+  fit_set = false;
+  needs_opt = false;
+  opt_rate = 0;
+  genetic_age = 0;
 }
 
-int get_arity(int node) {
-  if (AcyclicGraph::is_terminal(node)) return 0;
-  return AcyclicGraph::has_arity_two(node) ? 2 : 1;
+AcyclicGraph::AcyclicGraph(const AcyclicGraph &ag) {
+  stack = ag.stack;
+  constants = ag.constants;
+  simple_stack = ag.simple_stack;
+  fitness = ag.fitness;
+  fit_set = ag.fit_set;
+  needs_opt = ag.needs_opt;
+  opt_rate = ag.opt_rate;
+  genetic_age = ag.genetic_age;
 }
 
-
-
-Eigen::ArrayXXd SimplifyAndEvaluate(const Eigen::ArrayX3i & stack,
-                                    const Eigen::ArrayXXd & x,
-                                    const Eigen::VectorXd &constants) {
-  std::vector<bool> mask = FindUsedCommands(stack);
-  std::vector<Eigen::ArrayXXd> forward_eval = EvaluateWithMask(stack, x, constants, mask);
-  return forward_eval.back();
+AcyclicGraph AcyclicGraph::copy() {
+  AcyclicGraph temp = AcyclicGraph();
+  temp.stack = stack;
+  temp.constants = constants;
+  temp.simple_stack = simple_stack;
+  temp.fitness = fitness;
+  temp.fit_set = fit_set;
+  temp.needs_opt = needs_opt;
+  temp.opt_rate = opt_rate;
+  temp.genetic_age = genetic_age;
+  return temp;
 }
 
-std::pair<Eigen::ArrayXXd, Eigen::ArrayXXd> SimplifyAndEvaluateWithDerivative(
-  const Eigen::ArrayX3i &stack,
-  const Eigen::ArrayXXd &x,
-  const Eigen::VectorXd &constants,
-  const bool param_x_or_c) {
-  std::vector<bool> mask = FindUsedCommands(stack);
-  return EvaluateWithDerivativeAndMask(stack, x, constants, mask, param_x_or_c);
-}
+const char *AcyclicGraph::stack_print_map[13] = {
+  "X",
+  "C",
+  "+",
+  "-",
+  "*",
+  "/",
+  "sin",
+  "cos",
+  "exp",
+  "log",
+  "pow",
+  "abs",
+  "sqrt"
+};
 
-std::vector<bool> FindUsedCommands(const Eigen::ArrayX3i & stack) {
-  std::vector<bool> used_command(stack.rows());
-  used_command.back() = true;
-  for (int i = stack.rows() - 1; i >= 0; --i) {
-    if (used_command[i]) {
-      for (int j = 0; j < get_arity(stack(i, 0)); ++j) {
-        used_command[stack(i, j + 1)] = true;
+const bool AcyclicGraph::is_arity_2_map[13] = {
+  false,
+  false,
+  true,
+  true,
+  true,
+  true,
+  false,
+  false,
+  false,
+  false,
+  true,
+  false,
+  false
+};
+
+const bool AcyclicGraph::is_terminal_map[13] = {
+  true,
+  true,
+  false,
+  false,
+  false,
+  false,
+  false,
+  false,
+  false,
+  false,
+  false,
+  false,
+  false
+};
+
+bool AcyclicGraph::needs_optimization() {
+  if (opt_rate == 0) {
+    for (int i = 0; i < simple_stack.rows(); ++i) {
+      if (simple_stack(i, 0) == 1 && simple_stack(i, 1) == -1 ||
+          simple_stack(i, 0) == 1 && simple_stack(i, 1) >= constants.size()) {
+        return true;
       }
     }
+
+    return false;
+
+  } else {
+    return needs_opt;
   }
-  return used_command;
 }
 
-std::vector<Eigen::ArrayXXd> EvaluateWithMask(const Eigen::ArrayX3i &stack,
-                                              const Eigen::ArrayXXd &x,
-                                              const Eigen::VectorXd &constants,
-                                              const std::vector<bool> &mask) {
-  std::vector<Eigen::ArrayXXd> forward_eval(stack.rows());
+void AcyclicGraph::set_constants(Eigen::VectorXd con) {
+  constants = con;
+}
 
-  for (int i = 0; i < stack.rows(); ++i) {
-    if (mask[i]) {
-      int node = stack(i, NODE_IDX);
-      int op1 = stack(i, OP_1);
-      int op2 = stack(i, OP_2);
-      forward_eval[i] = backendnodes::forward_eval_function(
-        node, op1, op2, x, constants, forward_eval);
+int AcyclicGraph::count_constants() {
+  if (opt_rate == 0) {
+    std::set<int> util = utilized_commands();
+    std::set<int>::iterator it = util.begin();
+    int const_num = 0;
+
+    for (int i = 0; i < simple_stack.rows(); ++i, ++it) {
+      if (simple_stack(i, 0) == 1) {
+        simple_stack(i, 1) = const_num;
+        simple_stack(i, 2) = const_num;
+        stack(*it, 1) = const_num;
+        stack(*it, 2) = const_num;
+        const_num += 1;
+      }
     }
+
+    return const_num;
+
+  } else {
+    return constants.size();
   }
-  return forward_eval;
 }
 
-std::pair<Eigen::ArrayXXd, Eigen::ArrayXXd> EvaluateWithDerivativeAndMask(
-    const Eigen::ArrayX3i &stack,
-    const Eigen::ArrayXXd &x,
-    const Eigen::VectorXd &constants,
-    const std::vector<bool> &mask,
-    const bool param_x_or_c) {
-  std::vector<Eigen::ArrayXXd> forward_eval = EvaluateWithMask(
-    stack, x, constants, mask);
-
-  std::pair<int, int> deriv_shape;
-  int deriv_wrt_node;
-  if (param_x_or_c) {  // true = x
-    deriv_shape = std::make_pair(x.rows(), x.cols());
-    deriv_wrt_node = 0;
-  } else {  // false = c
-    deriv_shape = std::make_pair(x.rows(), constants.size());
-    deriv_wrt_node = 1;
-  }
-
-  Eigen::ArrayXXd derivative = reverse_eval_with_mask(
-    deriv_shape, deriv_wrt_node, forward_eval, stack, mask);
-  return std::make_pair(forward_eval.back(), derivative);
-}
-
-
-Eigen::ArrayXXd Evaluate(const Eigen::ArrayX3i & stack,
-                         const Eigen::ArrayXXd &x,
-                         const Eigen::VectorXd &constants) {
-  std::vector<bool> use_all_mask(stack.rows());
-  std::fill(use_all_mask.begin(), use_all_mask.end(), true);
-  std::vector<Eigen::ArrayXXd> forward_eval = EvaluateWithMask(
-    stack, x, constants, use_all_mask);
-  return forward_eval.back();  
-}
-
-std::pair<Eigen::ArrayXXd, Eigen::ArrayXXd> EvaluateWithDerivative(
-    const Eigen::ArrayX3i &stack,
-    const Eigen::ArrayXXd &x,
-    const Eigen::VectorXd &constants,
-    const bool param_x_or_c) {
-  std::vector<bool> use_all_mask(stack.rows());
-  std::fill(use_all_mask.begin(), use_all_mask.end(), true);
-  return EvaluateWithDerivativeAndMask(stack, x, constants, use_all_mask);
-}
-
-// void PrintStack(const Eigen::ArrayX3i & stack) {
-//   // Prints a stack to std::cout.
-//   for (std::size_t i = 0; i < stack.rows(); ++i) {
-//     //this is the operator
-//     std::cout << "(" << i << ") = " <<
-//               oper_interface.operator_map[stack(i, 0)]->get_print() << " : ";
-
-//     // Hard code for those with arity == 0
-//     if (get_arity(stack(i, 0)]->get_arity() == 0) {
-//       std::cout << " (" << stack(i, 1) << ")";
+// void AcyclicGraph::input_constants() {
+//   std::set<int> util = utilized_commands();
+//   std::set<int>::iterator it = util.begin();
+//   int const_num = 0;
+//   int j = 0;
+//   for (int i = 0; i < stack.rows(); ++i) {
+//     if (i == *it) {
+//       if (stack(i, 0) == 1) {
+//         simple_stack(j, 1) = const_num;
+//         simple_stack(j, 2) = const_num;
+//         stack(*it, 1) = const_num;
+//         stack(*it, 2) = const_num;
+//         const_num += 1;
+//       }
+//       ++it;
+//       ++j;
 //     }
-
-//     // loop through the rest dependent on their arity
-//     for (int j = 0; j < get_arity(stack(i, 0)); ++j) {
-//       std::cout << " (" << stack(i, j) << ")";
-//     }
-
-//     std::cout << std::endl;
+//     else
+//       if (stack(i, 0) == 1) {
+//         stack(i, 1) = -1;
+//         stack(i, 2) = -1;
+//       }
 //   }
+
+//   // if (const_num != count_constants()) {
+//   //   Eigen::VectorXd temp(const_num);
+//   //   constants = temp;
+//   // }
+//   if (const_num != count_constants())
+//     // std::cout << "Constants are too big\n";
+//     constants.conservativeResize(const_num);
 // }
 
-Eigen::ArrayX3i SimplifyStack(const Eigen::ArrayX3i & stack) {
-  std::vector<bool> used_command = FindUsedCommands(stack);
-  std::map<int, int> reduced_param_map;
-  Eigen::ArrayX3i new_stack(used_command.size(), 3);
-
-  for (int i = 0, j = 0; i < stack.rows(); ++i) {
-    if (used_command[i]) {
-      for (int k = 0; k < get_arity(new_stack(j, 0)); ++k) {
-        new_stack(j, k + 1) = reduced_param_map[new_stack(j, k + 1)];
-      }
-      reduced_param_map[i] = j;
-      ++j;
-    }
-  }
-  return new_stack;
+Eigen::ArrayXXd AcyclicGraph::evaluate(Eigen::ArrayXXd &eval_x) {
+  return Evaluate(simple_stack, eval_x, constants);
 }
 
-Eigen::ArrayXXd reverse_eval_with_mask(const std::pair<int, int> deriv_shape,
-                                       const int deriv_wrt_node,
-                                       const std::vector<Eigen::ArrayXXd> &forward_eval,
-                                       const Eigen::ArrayX3i &stack,
-                                       const std::vector<bool> &mask) {
-  int num_samples = deriv_shape.first;
-  int num_features = deriv_shape.second;
-  int stack_depth = stack.rows();
-  Eigen::ArrayXXd zero = Eigen::ArrayXd::Zero(num_samples);
-  Eigen::ArrayXXd ones = Eigen::ArrayXd::Ones(num_samples);
+std::pair<Eigen::ArrayXXd, Eigen::ArrayXXd> AcyclicGraph::evaluate_deriv(
+  Eigen::ArrayXXd &eval_x) {
+  return EvaluateWithDerivative(simple_stack, eval_x, constants);
+}
 
-  Eigen::ArrayXXd derivative = Eigen::ArrayXXd::Zero(num_samples, num_features);
-  std::vector<Eigen::ArrayXXd> reverse_eval(stack_depth); 
-  std::fill(reverse_eval.begin(), reverse_eval.end(), zero); 
-  reverse_eval[stack_depth-1] = ones;
+std::pair<Eigen::ArrayXXd, Eigen::ArrayXXd>
+AcyclicGraph::evaluate_with_const_deriv(
+  Eigen::ArrayXXd &eval_x) {
+  return EvaluateWithDerivative(simple_stack, eval_x, constants, false);
+}
 
-  for (int i = stack_depth - 1; i >= 0; i--) {
-    if (mask[i]) {
-      int node = stack(i, NODE_IDX);
-      int param1 = stack(i, OP_1);
-      int param2 = stack(i, OP_2);
-      if (node == deriv_wrt_node)  {
-        derivative.col(param1) += reverse_eval[i];
-      } else {
-        backendnodes::reverse_eval_function(node, i, param1, param2,
-                                            forward_eval,
-                                            reverse_eval);
-      }
+
+// TODO: remove iterator interface
+std::string AcyclicGraph::latexstring() {
+  std::vector<std::string> strings;
+  std::ostringstream stream;
+
+  for (int i = 0; i < simple_stack.rows(); ++i) {
+    std::string temp = "";
+
+    switch ((int)simple_stack(i, 0)) {
+      case 0:
+        stream << oper_interface.operator_map[simple_stack(i, 0)]->get_print()
+               << "_" << simple_stack(i, 1);
+        break;
+
+      case 1:
+        if (simple_stack(i, 1) == -1) {
+          stream << "0";
+
+        } else {
+          stream << constants[simple_stack(i, 1)];
+        }
+
+        break;
+
+      case 2:
+        stream << strings[simple_stack(i, 1)] << " + " <<
+               strings[simple_stack(i, 2)];
+        break;
+
+      case 3:
+        stream << strings[simple_stack(i, 1)] << " - (" <<
+               strings[simple_stack(i, 2)] << ")";
+        break;
+
+      case 4:
+        stream << "(" << strings[simple_stack(i, 1)] << ")(" <<
+               strings[simple_stack(i, 2)] << ")";
+        break;
+
+      case 5:
+        stream << "\\frac{" << strings[simple_stack(i, 1)] << "}{" <<
+               strings[simple_stack(i, 2)] << "}";
+        break;
+
+      case 6:
+      case 7:
+      case 8:
+      case 9:
+      case 12:
+        stream << "\\" << oper_interface.operator_map[simple_stack(i, 0)]
+               ->get_print() << "{" << strings[simple_stack(i, 1)] << "}";
+        break;
+
+      case 10:
+        stream << "(" << strings[simple_stack(i, 1)] << ")^{(" <<
+               strings[simple_stack(i, 2)] << ")}";
+        break;
+
+      case 11:
+        stream << "|{" << strings[simple_stack(i, 1)] << "}|";
+        break;
+
+      default:
+        stream << "";
+        break;
+    }
+
+    temp = stream.str();
+    strings.push_back(temp);
+    stream.str(std::string());
+  }
+
+  return strings.back();
+}
+
+std::set<int> AcyclicGraph::utilized_commands() {
+  std::set<int> util;
+  util.insert(stack.rows() - 1);
+
+  for (int i = stack.rows() - 1; i >= 0; --i) {
+    if (stack(i, 0) > 1 && util.count(i) == 1) {
+      util.insert(stack(i, 1));
+      util.insert(stack(i, 2));
     }
   }
-  return derivative;
+
+  return util;
+}
+
+int AcyclicGraph::complexity() {
+  return simple_stack.rows();
+}
+
+std::string AcyclicGraph::print_stack() {
+  std::ostringstream out;
+  out << "---full stack---\n";
+
+  for (int i = 0; i < stack.rows(); ++i) {
+    out << std::left << std::setw(4) << i;
+    out << "<= ";
+
+    if (stack(i, 0) == 0)
+      out << oper_interface.operator_map[stack(i, 0)]->get_print()
+          << stack(i, 1) << std::endl;
+
+    else if (stack(i, 0) == 1) {
+      if (stack(i, 1) == -1) {
+        out << oper_interface.operator_map[stack(i, 0)]->get_print();
+
+      } else {
+        out << constants[stack(i, 1)];
+      }
+
+      out << std::endl;
+
+    } else {
+      out << "(" << stack(i, 1) << ") "
+          << oper_interface.operator_map[stack(i, 0)]->get_print()
+          << " (" << stack(i, 2) << ")\n";
+    }
+  }
+
+  out << "---small stack---\n";
+
+  for (int i = 0; i < simple_stack.rows(); ++i) {
+    out << std::left << std::setw(4) << i;
+    out << "<= ";
+
+    if (simple_stack(i, 0) == 0)
+      out << oper_interface.operator_map[simple_stack(i, 0)]->get_print()
+          << simple_stack(i, 1) << std::endl;
+
+    else if (simple_stack(i, 0) == 1) {
+      if (simple_stack(i, 1) == -1) {
+        out << oper_interface.operator_map[simple_stack(i, 0)]->get_print();
+
+      } else {
+        out << constants[simple_stack(i, 1)];
+      }
+
+      out << std::endl;
+
+    } else {
+      out << "(" << simple_stack(i, 1) << ") "
+          << oper_interface.operator_map[simple_stack(i, 0)]->get_print()
+          << " (" << simple_stack(i, 2) << ")\n";
+    }
+  }
+
+  return out.str();
+}
+
+bool AcyclicGraph::has_arity_two(int node) {
+  return is_arity_2_map[node];
+}
+
+bool AcyclicGraph::is_terminal(int node) {
+  return is_terminal_map[node];
 }
